arch: pay down all arch_guard overages and gate the full guard in CI (#15643)
Fixes every live arch_guard failure on main and wires the full guard into
the arch-size CI job so guard drift can no longer land silently:

- #8225: drain query_boundaries/common.rs 1945 -> 1740 lines by moving
  FunctionShape instantiation helpers to generic_instantiation, param-list
  transformation helpers to signature_building, and redeclaration-widening
  helpers to widening; delete dead enum_or_lazy_def_id; ratchet the LOC
  baseline 1901 -> 1740 and the reference baseline 3082 -> 3050.
- #8205: split lib-builtin registry accessors onto a narrower
  TypeBuiltinAccess supertrait of TypeDatabase (81 -> 75 methods, cap 75).
- Metric 7: route tsz-lsp completions traversal_state through the module
  chain like its siblings (36 -> 35 importing files).
- Track 10: consolidate fifteen cross-arena delegation sites onto a single
  CheckerState::delegate_for_arena factory (with_parent_cache_attributed
  29 -> 16 callsites, copy_symbol_file_targets_to_attributed 19 -> 5);
  split state/delegation.rs out of the oversized state.rs.
- Solver: replace bare .unwrap() with .expect() in constraints test
  modules; use interner.lazy() instead of raw intern(TypeData::Lazy).
- ObjectFlags rule: sync the stale TOML exclude list with the canonical
  per-surface boundary set in object_flags_boundary_scans.rs (#14351) and
  add a policy self-test that keeps the two lists in sync.
- Delete scripts/arch/arch_guard_config.py: a dead, drifted duplicate of
  arch_guard_shared.py that nothing imports.
- Pin the five production files that crossed 2000 lines while the guard
  was ungated (type_queries/core.rs, transforms/helpers.rs,
  class_es5_ir_members.rs, emitter/statements/core.rs, subtype/explain.rs).
- Fix the three arch scan tests that fail on clean main: route the
  generator yield union through function_returns, repoint the moved
  unrelated_type_parameter_target_related_info scan, and add the
  unknown_satisfies_top_type_constraint named outcome helper for the
  TS2344 unknown-constraint check in constraint_validation.rs.
- CI: arch-size job now runs the full guard plus its 468 self-tests on
  merge_group/push/dispatch.
