+++ .github/workflows/ci.yml
@@ -229,13 +229,22 @@ jobs:
     name: arch-size
     if: ${{ github.event_name != 'pull_request' }}
     runs-on: ubuntu-latest
-    timeout-minutes: 5
+    timeout-minutes: 10
     steps:
       - uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7.0.0
         with:
           fetch-depth: 1
-      - name: Run architecture size guard (LOC caps + non-#8225 size ratchets)
-        run: python3 scripts/arch/arch_guard.py --size-only
+      # Full guard (#15643): size caps, #8225 quarantine ratchets, metric-7
+      # import counts, Track 10 residency counts, and boundary pattern checks
+      # all gate merges. Ratchet baselines live in
+      # scripts/arch/arch_guard_shared.py; pay down or intentionally bump them
+      # in the same PR that moves a count.
+      - name: Run full architecture guard
+        run: python3 scripts/arch/arch_guard.py
+      - name: Run architecture guard self-tests
+        run: |
+          cd scripts/arch
+          python3 -m unittest discover -p "test_arch_guard*.py" -v 2>&1 | tail -20
 
   ci-summary:
     name: CI Summary
