+++ crates/tsz-checker/src/jsdoc/resolution/type_construction.rs
@@ -63,22 +63,14 @@ impl<'a> CheckerState<'a> {
         let binder = all_binders.get(file_idx)?;
         let source_file = arena.source_files.first()?;
 
-        let mut checker = Box::new(CheckerState::with_parent_cache_attributed(
+        let mut checker = CheckerState::delegate_for_arena(
             arena.as_ref(),
             binder.as_ref(),
-            self.ctx.types,
             source_file.file_name.clone(),
-            self.ctx.compiler_options.clone(),
             self,
             tsz_common::perf_counters::CheckerCreationReason::JsDocTypeConstruction,
-        ));
-        checker.ctx.lib_contexts = self.ctx.lib_contexts.clone();
-        checker.ctx.copy_cross_file_state_from(&self.ctx);
-        checker.ctx.current_file_idx = file_idx;
-        self.ctx.copy_symbol_file_targets_to_attributed(
-            &mut checker.ctx,
-            tsz_common::perf_counters::CheckerCreationReason::JsDocTypeConstruction,
         );
+        checker.ctx.current_file_idx = file_idx;
 
         let result = checker.jsdoc_enum_annotation_type_for_current_checker(decl);
         self.ctx.merge_symbol_file_targets_from(&checker.ctx);
@@ -181,22 +173,14 @@ impl<'a> CheckerState<'a> {
             }
 
             for source_file in &arena.source_files {
-                let mut checker = Box::new(CheckerState::with_parent_cache_attributed(
+                let mut checker = CheckerState::delegate_for_arena(
                     arena.as_ref(),
                     binder.as_ref(),
-                    self.ctx.types,
                     source_file.file_name.clone(),
-                    self.ctx.compiler_options.clone(),
                     self,
                     tsz_common::perf_counters::CheckerCreationReason::JsDocTypeConstruction,
-                ));
-                checker.ctx.lib_contexts = self.ctx.lib_contexts.clone();
-                checker.ctx.copy_cross_file_state_from(&self.ctx);
-                checker.ctx.current_file_idx = file_idx;
-                self.ctx.copy_symbol_file_targets_to_attributed(
-                    &mut checker.ctx,
-                    tsz_common::perf_counters::CheckerCreationReason::JsDocTypeConstruction,
                 );
+                checker.ctx.current_file_idx = file_idx;
 
                 if let Some(ty) = checker.resolve_jsdoc_assigned_value_type_in_arena(
                     name,
