+++ crates/tsz-checker/src/state/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod cache_invalidation;
+pub(crate) mod delegation;
 pub(crate) mod request_cache;
 pub mod state;
 pub mod state_checking;
