+++ crates/tsz-checker/src/state/state.rs
@@ -310,81 +310,6 @@ impl<'a> CheckerState<'a> {
         }
     }
 
-    /// Create a child `CheckerState` that shares the parent's caches.
-    /// This is used for temporary checkers (e.g., cross-file symbol resolution)
-    /// to ensure cache results are not lost (fixes Cache Isolation Bug).
-    pub fn with_parent_cache(
-        arena: &'a NodeArena,
-        binder: &'a BinderState,
-        types: &'a dyn QueryDatabase,
-        file_name: String,
-        compiler_options: CheckerOptions,
-        parent: &Self,
-    ) -> Self {
-        // Attribution: prefer `with_parent_cache_attributed` at call sites
-        // we want to track in the per-reason counter dump (PR #1631).
-        // Sites that still call this raw form attribute to
-        // `CheckerCreationReason::Other`. See
-        // `docs/plan/PERFORMANCE_PLAN.md`.
-        Self::with_parent_cache_attributed(
-            arena,
-            binder,
-            types,
-            file_name,
-            compiler_options,
-            parent,
-            tsz_common::perf_counters::CheckerCreationReason::Other,
-        )
-    }
-
-    /// Attributed variant of [`Self::with_parent_cache`]: the caller passes
-    /// the reason this child checker is being created so PR #1631's
-    /// counter dump can show which call sites drive the construction
-    /// explosion. Always prefer this over the raw `with_parent_cache`.
-    pub fn with_parent_cache_attributed(
-        arena: &'a NodeArena,
-        binder: &'a BinderState,
-        types: &'a dyn QueryDatabase,
-        file_name: String,
-        compiler_options: CheckerOptions,
-        parent: &Self,
-        reason: tsz_common::perf_counters::CheckerCreationReason,
-    ) -> Self {
-        tsz_common::perf_counters::record_with_parent_cache(reason);
-        CheckerState {
-            ctx: CheckerContext::with_parent_cache(
-                arena,
-                binder,
-                types,
-                file_name,
-                compiler_options,
-                &parent.ctx,
-            ),
-        }
-    }
-
-    /// Copy resolution-cycle guard sets from `parent` into this child checker and
-    /// wire up the shared `DefinitionStore`. Call immediately after constructing a
-    /// class-delegation child checker via `with_parent_cache_attributed`.
-    pub(super) fn propagate_class_delegation_setup(
-        &mut self,
-        parent: &CheckerState,
-        skip_sym: tsz_binder::SymbolId,
-    ) {
-        for &id in &parent.ctx.class_instance_resolution_set {
-            self.ctx.class_instance_resolution_set.insert(id);
-        }
-        for &id in &parent.ctx.symbol_resolution_set {
-            if id != skip_sym {
-                self.ctx.symbol_resolution_set.insert(id);
-            }
-        }
-        for &id in &parent.ctx.class_constructor_resolution_set {
-            self.ctx.class_constructor_resolution_set.insert(id);
-        }
-        self.ctx.ensure_both_envs_have_definition_store();
-    }
-
     /// Thread-local guard for cross-arena delegation depth.
     /// All cross-arena delegation points (`delegate_cross_arena_symbol_resolution`,
     /// `get_type_params_for_symbol`, `type_of_value_declaration`) MUST call this
