+++ crates/tsz-checker/src/state/type_analysis/computed_alias.rs
@@ -46,19 +46,11 @@ impl CheckerState<'_> {
             .map(|sf| sf.file_name.clone())
             .unwrap_or_else(|| self.ctx.file_name.clone());
 
-        let mut checker = Box::new(CheckerState::with_parent_cache_attributed(
+        let mut checker = CheckerState::delegate_for_arena(
             decl_arena,
             delegate_binder,
-            self.ctx.types,
             delegate_file_name,
-            self.ctx.compiler_options.clone(),
             self,
-            tsz_common::perf_counters::CheckerCreationReason::AliasResolution,
-        ));
-        checker.ctx.lib_contexts = self.ctx.lib_contexts.clone();
-        checker.ctx.copy_cross_file_state_from(&self.ctx);
-        self.ctx.copy_symbol_file_targets_to_attributed(
-            &mut checker.ctx,
             tsz_common::perf_counters::CheckerCreationReason::AliasResolution,
         );
         checker.ctx.current_file_idx = delegate_file_idx.unwrap_or(self.ctx.current_file_idx);
