+++ crates/tsz-checker/src/state/type_analysis/computed_commonjs/child_checker.rs
@@ -40,22 +40,14 @@ impl<'a> CheckerState<'a> {
         let binder = all_binders.get(target_file_idx)?;
         let source_file = arena.source_files.first()?;
 
-        let mut checker = Box::new(CheckerState::with_parent_cache_attributed(
+        let mut checker = CheckerState::delegate_for_arena(
             arena.as_ref(),
             binder.as_ref(),
-            self.ctx.types,
             source_file.file_name.clone(),
-            self.ctx.compiler_options.clone(),
             self,
             CheckerCreationReason::CjsExports,
-        ));
-        checker.ctx.lib_contexts = self.ctx.lib_contexts.clone();
-        checker.ctx.copy_cross_file_state_from(&self.ctx);
-        checker.ctx.current_file_idx = target_file_idx;
-        self.ctx.copy_symbol_file_targets_to_attributed(
-            &mut checker.ctx,
-            CheckerCreationReason::CjsExports,
         );
+        checker.ctx.current_file_idx = target_file_idx;
 
         let result = f(&mut checker);
         if merge_symbol_file_targets {
