+++ crates/tsz-checker/src/state/type_analysis/cross_file.rs
@@ -1549,19 +1549,11 @@ impl CheckerState<'_> {
         tsz_common::perf_counters::record_delegate_cross_arena_miss();
         let _delegate_depth_guard = tsz_common::perf_counters::enter_delegate();
 
-        let mut checker = Box::new(CheckerState::with_parent_cache_attributed(
+        let mut checker = CheckerState::delegate_for_arena(
             symbol_arena,
             delegate_binder,
-            self.ctx.types,
             delegate_file_name,
-            self.ctx.compiler_options.clone(),
             self,
-            tsz_common::perf_counters::CheckerCreationReason::DelegateCrossArenaInterface,
-        ));
-        checker.ctx.lib_contexts = self.ctx.lib_contexts.clone();
-        checker.ctx.copy_cross_file_state_from(&self.ctx);
-        self.ctx.copy_symbol_file_targets_to_attributed(
-            &mut checker.ctx,
             tsz_common::perf_counters::CheckerCreationReason::DelegateCrossArenaInterface,
         );
         checker.ctx.current_file_idx = delegate_file_idx.unwrap_or(self.ctx.current_file_idx);
@@ -1777,19 +1769,11 @@ impl CheckerState<'_> {
         tsz_common::perf_counters::record_delegate_cross_arena_miss();
         let _delegate_depth_guard = tsz_common::perf_counters::enter_delegate();
 
-        let mut checker = Box::new(CheckerState::with_parent_cache_attributed(
+        let mut checker = CheckerState::delegate_for_arena(
             interface_arena,
             delegate_binder,
-            self.ctx.types,
             delegate_file_name,
-            self.ctx.compiler_options.clone(),
             self,
-            tsz_common::perf_counters::CheckerCreationReason::DelegateCrossArenaOther,
-        ));
-        checker.ctx.lib_contexts = self.ctx.lib_contexts.clone();
-        checker.ctx.copy_cross_file_state_from(&self.ctx);
-        self.ctx.copy_symbol_file_targets_to_attributed(
-            &mut checker.ctx,
             tsz_common::perf_counters::CheckerCreationReason::DelegateCrossArenaOther,
         );
         checker.ctx.current_file_idx = delegate_file_idx.unwrap_or(self.ctx.current_file_idx);
