+++ crates/tsz-checker/src/state/type_environment/core.rs
@@ -53,19 +53,11 @@ impl CheckerState<'_> {
                 .unwrap_or_else(|| self.ctx.file_name.clone());
             tsz_common::perf_counters::record_delegate_cross_arena_miss();
             _enum_decl_delegate_guard = Some(tsz_common::perf_counters::enter_delegate());
-            let mut checker = Box::new(CheckerState::with_parent_cache_attributed(
+            let mut checker = CheckerState::delegate_for_arena(
                 enum_arena,
                 enum_binder,
-                self.ctx.types,
                 file_name,
-                self.ctx.compiler_options.clone(),
                 self,
-                tsz_common::perf_counters::CheckerCreationReason::TypeEnvironmentCore,
-            ));
-            checker.ctx.lib_contexts = self.ctx.lib_contexts.clone();
-            checker.ctx.copy_cross_file_state_from(&self.ctx);
-            self.ctx.copy_symbol_file_targets_to_attributed(
-                &mut checker.ctx,
                 tsz_common::perf_counters::CheckerCreationReason::TypeEnvironmentCore,
             );
             checker.ctx.current_file_idx = file_idx;
