+++ crates/tsz-checker/src/state/type_resolution/cross_file_constructors.rs
@@ -38,19 +38,11 @@ impl<'a> CheckerState<'a> {
             .first()
             .map(|sf| sf.file_name.clone())
             .unwrap_or_else(|| self.ctx.file_name.clone());
-        let mut checker = Box::new(CheckerState::with_parent_cache_attributed(
+        let mut checker = CheckerState::delegate_for_arena(
             arena.as_ref(),
             binder,
-            self.ctx.types,
             file_name,
-            self.ctx.compiler_options.clone(),
             self,
-            CheckerCreationReason::DelegateCrossArenaOther,
-        ));
-        checker.ctx.lib_contexts = self.ctx.lib_contexts.clone();
-        checker.ctx.copy_cross_file_state_from(&self.ctx);
-        self.ctx.copy_symbol_file_targets_to_attributed(
-            &mut checker.ctx,
             CheckerCreationReason::DelegateCrossArenaOther,
         );
         checker.ctx.current_file_idx = file_idx;
