+++ crates/tsz-checker/src/state/type_resolution/import_type.rs
@@ -477,22 +477,14 @@ impl<'a> CheckerState<'a> {
             // No cache fast-path on this delegate; every entry is a miss.
             tsz_common::perf_counters::record_delegate_cross_arena_miss();
             let _delegate_depth_guard = tsz_common::perf_counters::enter_delegate();
-            let mut checker = Box::new(CheckerState::with_parent_cache_attributed(
+            let mut checker = CheckerState::delegate_for_arena(
                 &target_arena,
                 &target_binder,
-                self.ctx.types,
                 source_file.file_name.clone(),
-                self.ctx.compiler_options.clone(),
                 self,
                 tsz_common::perf_counters::CheckerCreationReason::ImportType,
-            ));
-            checker.ctx.lib_contexts = self.ctx.lib_contexts.clone();
-            checker.ctx.copy_cross_file_state_from(&self.ctx);
-            checker.ctx.current_file_idx = target_file_idx;
-            self.ctx.copy_symbol_file_targets_to_attributed(
-                &mut checker.ctx,
-                tsz_common::perf_counters::CheckerCreationReason::ImportType,
             );
+            checker.ctx.current_file_idx = target_file_idx;
 
             if let Some((ty, _)) =
                 checker.resolve_jsdoc_typedef_info(typedef_name, &comments, &source_text)
@@ -543,22 +535,14 @@ impl<'a> CheckerState<'a> {
             // No cache fast-path on this delegate; every entry is a miss.
             tsz_common::perf_counters::record_delegate_cross_arena_miss();
             let _delegate_depth_guard = tsz_common::perf_counters::enter_delegate();
-            let mut checker = Box::new(CheckerState::with_parent_cache_attributed(
+            let mut checker = CheckerState::delegate_for_arena(
                 &target_arena,
                 &target_binder,
-                self.ctx.types,
                 source_file.file_name.clone(),
-                self.ctx.compiler_options.clone(),
                 self,
                 tsz_common::perf_counters::CheckerCreationReason::ImportType,
-            ));
-            checker.ctx.lib_contexts = self.ctx.lib_contexts.clone();
-            checker.ctx.copy_cross_file_state_from(&self.ctx);
-            checker.ctx.current_file_idx = target_file_idx;
-            self.ctx.copy_symbol_file_targets_to_attributed(
-                &mut checker.ctx,
-                tsz_common::perf_counters::CheckerCreationReason::ImportType,
             );
+            checker.ctx.current_file_idx = target_file_idx;
 
             if let Some((_, type_params)) =
                 checker.resolve_jsdoc_typedef_info(member_name, &comments, &source_text)
