+++ crates/tsz-checker/src/types/module_augmentation_value.rs
@@ -55,19 +55,11 @@ impl<'a> CheckerState<'a> {
         tsz_common::perf_counters::record_delegate_cross_arena_miss();
         let _delegate_depth_guard = tsz_common::perf_counters::enter_delegate();
 
-        let mut checker = Box::new(CheckerState::with_parent_cache_attributed(
+        let mut checker = CheckerState::delegate_for_arena(
             arena,
             delegate_binder,
-            self.ctx.types,
             delegate_file_name,
-            self.ctx.compiler_options.clone(),
             self,
-            CheckerCreationReason::ModuleAugmentationValue,
-        ));
-        checker.ctx.lib_contexts = self.ctx.lib_contexts.clone();
-        checker.ctx.copy_cross_file_state_from(&self.ctx);
-        self.ctx.copy_symbol_file_targets_to_attributed(
-            &mut checker.ctx,
             CheckerCreationReason::ModuleAugmentationValue,
         );
         checker.ctx.current_file_idx = delegate_file_idx.unwrap_or(self.ctx.current_file_idx);
