+++ crates/tsz-checker/src/types/property_access_helpers/expando.rs
@@ -394,19 +394,11 @@ impl<'a> CheckerState<'a> {
         tsz_common::perf_counters::record_delegate_cross_arena_miss();
         let _delegate_depth_guard = tsz_common::perf_counters::enter_delegate();
 
-        let mut checker = Box::new(CheckerState::with_parent_cache_attributed(
+        let mut checker = CheckerState::delegate_for_arena(
             arena,
             binder,
-            self.ctx.types,
             file_name,
-            self.ctx.compiler_options.clone(),
             self,
-            tsz_common::perf_counters::CheckerCreationReason::ExpandoProperty,
-        ));
-        checker.ctx.lib_contexts = self.ctx.lib_contexts.clone();
-        checker.ctx.copy_cross_file_state_from(&self.ctx);
-        self.ctx.copy_symbol_file_targets_to_attributed(
-            &mut checker.ctx,
             tsz_common::perf_counters::CheckerCreationReason::ExpandoProperty,
         );
         checker.ctx.current_file_idx = file_idx;
