+++ crates/tsz-checker/src/types/queries/callable_truthiness.rs
@@ -338,19 +338,11 @@ impl<'a> CheckerState<'a> {
             // No cache fast-path on this delegate; every entry is a miss.
             tsz_common::perf_counters::record_delegate_cross_arena_miss();
             let _delegate_depth_guard = tsz_common::perf_counters::enter_delegate();
-            let mut checker = Box::new(CheckerState::with_parent_cache_attributed(
+            let mut checker = CheckerState::delegate_for_arena(
                 arena,
                 delegate_binder,
-                self.ctx.types,
                 delegate_file_name,
-                self.ctx.compiler_options.clone(),
                 self,
-                tsz_common::perf_counters::CheckerCreationReason::CallableTruthiness,
-            ));
-            checker.ctx.lib_contexts = self.ctx.lib_contexts.clone();
-            checker.ctx.copy_cross_file_state_from(&self.ctx);
-            self.ctx.copy_symbol_file_targets_to_attributed(
-                &mut checker.ctx,
                 tsz_common::perf_counters::CheckerCreationReason::CallableTruthiness,
             );
             checker.ctx.current_file_idx = delegate_file_idx;
