+++ crates/tsz-solver/src/lib.rs
@@ -248,7 +248,9 @@ pub mod observability {
 /// These create or modify types via the interner. Should be accessed through
 /// `query_boundaries` in the checker crate.
 pub mod construction {
-    pub use crate::caches::db::{QueryDatabase, TypeDatabase, TypeSubstitutionConstruction};
+    pub use crate::caches::db::{
+        QueryDatabase, TypeBuiltinAccess, TypeDatabase, TypeSubstitutionConstruction,
+    };
     pub use crate::caches::query_cache::{QueryCache, RelationCacheProbe, SharedQueryCache};
     pub use crate::caches::query_cache_statistics::{QueryCacheStatistics, RelationCacheStats};
     pub use crate::intern::type_factory::*;
