+++ docs/how-tsz-works/file-inventory/scripts-architecture-and-agents.md
@@ -10,7 +10,6 @@ Generated by `scripts/docs/generate-how-tsz-inventory.mjs`. This page mentions 3
 | `scripts/agents/test_disk_preflight.py` | Repository automation, tooling, harness, or application source. |
 | `scripts/agents/test_llm_context_audit.py` | Repository automation, tooling, harness, or application source. |
 | `scripts/arch/arch_guard.py` | Architecture guard, ratchet, or source-boundary tooling. |
-| `scripts/arch/arch_guard_config.py` | Architecture guard, ratchet, or source-boundary tooling. |
 | `scripts/arch/arch_guard_counts.py` | Architecture guard, ratchet, or source-boundary tooling. |
 | `scripts/arch/arch_guard_policy.toml` | Architecture guard, ratchet, or source-boundary tooling. |
 | `scripts/arch/arch_guard_project.py` | Architecture guard, ratchet, or source-boundary tooling. |
