--- scripts/arch/arch_guard_config.py
@@ -1,1067 +1,0 @@
-import pathlib
-import re
-
-ROOT = pathlib.Path(__file__).resolve().parents[2]
-
-# Pin field counts on giant coordination structs so workstream-4 (Checker
-# State / Speculation) extraction work shows up as visible metric drift in
-# the diff.  Each entry: (description, file_path, struct_name, max_fields).
-#
-# When a field is added intentionally, bump the cap in the same PR. This makes
-# architecture health metric drift visible at review time.
-STRUCT_FIELD_COUNT_CHECKS = [
-    (
-        "Checker boundary: CheckerContext field count (architecture health metric 1)",
-        ROOT / "crates" / "tsz-checker" / "src" / "context" / "mod.rs",
-        "CheckerContext",
-        253,
-    ),
-]
-
-# Pin the size of the solver's full database capability trait while #8205
-# splits it into narrower storage/config/provenance traits.  The live count is
-# tolerated as baseline debt, but new methods must either land on a narrower
-# trait or deliberately bump this cap with a roadmap/issue explanation.
-#
-# Each entry: (description, file_path, trait_name, max_methods).
-TRAIT_METHOD_COUNT_CHECKS = [
-    (
-        "Solver boundary: TypeDatabase method count (#8205)",
-        ROOT / "crates" / "tsz-solver" / "src" / "caches" / "db.rs",
-        "TypeDatabase",
-        80,
-    ),
-]
-
-VALID_CHECKER_CONTEXT_LIFETIMES = {
-    "ProgramStable",
-    "WorkerReusable",
-    "FileLocalReset",
-    "SpeculationScoped",
-    "DiagnosticsOnly",
-    "LspPersistent",
-}
-
-VALID_CHECKER_CONTEXT_CAPABILITIES = {
-    "CheckerInputs",
-    "DiagnosticState",
-    "EmitSummaryState",
-    "FileTypeCache",
-    "FlowSessionState",
-    "ProgramLookupContext",
-    "RelationSessionState",
-    "SpeculationState",
-}
-
-CHECKER_CONTEXT_LIFETIME_MANIFEST_CHECKS = [
-    (
-        "Checker boundary: CheckerContext lifetime inventory (T2.1.A)",
-        ROOT / "crates" / "tsz-checker" / "src" / "context" / "mod.rs",
-        "CheckerContext",
-        ROOT
-        / "crates"
-        / "tsz-checker"
-        / "src"
-        / "context"
-        / "checker_context_lifetimes.toml",
-    ),
-]
-
-# Pin the count of files that construct full independent parse→bind→check
-# pipelines (architecture health metric 4 in `docs/plan/ROADMAP.md`).  A
-# "full pipeline" is any non-test source file that calls all three of
-# `ParserState::new`, `BinderState::new`, `CheckerState::new` — that is, a
-# frontend reaching past the compiler service into the raw crate APIs.
-#
-# Workstream 3 ("Compiler Service Front Door") exit criterion is "There is
-# one blessed parse-bind-check path."  Pinning the count makes new
-# independent pipelines fail pre-commit and consolidation work show up as
-# a cap reduction in the same diff.
-#
-# Each entry: (description, search_roots, max_pipelines).
-INDEPENDENT_PIPELINE_CHECKS = [
-    (
-        "Frontend boundary: independent parse-bind-check pipelines (architecture health metric 4)",
-        [
-            ROOT / "crates" / "tsz-cli" / "src",
-            ROOT / "crates" / "tsz-core" / "src",
-            ROOT / "crates" / "tsz-lsp" / "src",
-            ROOT / "crates" / "tsz-wasm" / "src",
-        ],
-        4,
-    ),
-]
-
-# Pin the count of non-test source files that import `tsz_solver` outside the
-# solver/checker boundary (architecture health metric 7 in
-# `docs/plan/ROADMAP.md`).  The checker crate contains the canonical
-# `query_boundaries` modules and is the one architecturally allowed consumer
-# of solver internals; every other crate (`tsz-cli`, `tsz-core`, `tsz-lsp`,
-# `tsz-wasm`, `tsz-emitter`, `tsz-lowering`) reaching directly into the solver
-# weakens the front door story (workstream 3) and shows up as drift on this
-# metric.
-#
-# A file "imports tsz_solver" if a non-comment line contains one of:
-#   - `use tsz_solver::...`
-#   - `pub use tsz_solver` (re-export, including `pub use tsz_solver;`)
-#   - `extern crate tsz_solver`
-#
-# Each entry: (description, search_roots, exclude_path_prefixes, max_imports).
-SOLVER_IMPORT_COUNT_CHECKS = [
-    (
-        "Frontend/emitter boundary: direct tsz_solver imports outside solver/checker (architecture health metric 7)",
-        [ROOT / "crates"],
-        (
-            "crates/tsz-solver/",
-            "crates/tsz-checker/",
-        ),
-        36,
-    ),
-]
-
-# Pin the count of flat root-level solver computation API references outside
-# the approved checker query-boundary layer. Existing references are
-# transitional compatibility debt from `tsz_solver::*` root re-exports; new
-# references should go through a named solver facade, a checker
-# `query_boundaries` helper, or intentionally bump this cap.
-#
-# Each entry:
-#   (description, search_roots, exclude_path_prefixes, max_references).
-ROOT_SOLVER_COMPUTATION_IMPORT_COUNT_CHECKS = [
-    (
-        "Solver API boundary: flat root computation imports outside query boundaries (#8204)",
-        [
-            ROOT / "crates" / "tsz-checker" / "src",
-            ROOT / "crates" / "tsz-emitter" / "src",
-            ROOT / "crates" / "tsz-lsp" / "src",
-            ROOT / "crates" / "tsz-cli" / "src",
-        ],
-        ("crates/tsz-checker/src/query_boundaries/",),
-        0,
-    ),
-]
-
-# Seal the module-path escape hatch around the #8204 tiered solver API: the
-# flat-root ratchet above pins re-exported symbol names, but the solver still
-# declares `pub mod operations`/`relations`/`evaluation`/..., so downstream
-# crates can reach computation internals by full module path
-# (`tsz_solver::operations::widening::widen_type`). This pins those references
-# in the emitter/LSP/CLI/wasm source trees at the current count; migrating a
-# site to a tiered facade or checker query-boundary helper should ratchet the
-# cap down in the same diff.
-#
-# Transitional exceptions pinned by the current cap (#8204 successor debt —
-# each file below still reaches computation modules by path and is the
-# migration backlog for sealing the tiered API):
-#   - crates/tsz-cli/src/bin/tsz.rs (relations::subtype thread-local reset)
-#   - crates/tsz-cli/src/driver/check_tests/check_tests_part2.rs
-#     (operations::property::PropertyAccessResult)
-#   - crates/tsz-emitter/src/declaration_emitter/core/emit_class_properties.rs
-#   - crates/tsz-emitter/src/declaration_emitter/helpers/generic_call_literal.rs
-#   - crates/tsz-emitter/src/declaration_emitter/helpers/type_inference.rs
-#   - crates/tsz-emitter/src/declaration_emitter/helpers/
-#     type_inference_return_normalization.rs
-#   - crates/tsz-emitter/src/declaration_emitter/helpers/variable_decl.rs
-#   - crates/tsz-emitter/src/declaration_emitter/helpers/
-#     variable_decl_type_helpers.rs
-#     (operations::widening / operations iterator helpers)
-#   - crates/tsz-emitter/src/lowering/helpers_private_fields.rs
-#     (operations::compound_assignment)
-#   - crates/tsz-lsp/src/code_actions/code_action_extract.rs
-#     (operations::compound_assignment)
-#   - crates/tsz-lsp/src/completions/mod.rs (objects apparent-member helpers)
-#
-# Each entry:
-#   (description, search_roots, exclude_path_prefixes, max_references).
-MODULE_PATH_SOLVER_COMPUTATION_IMPORT_COUNT_CHECKS = [
-    (
-        "Solver API boundary: module-path computation imports in emitter/LSP/CLI/wasm (#8204)",
-        [
-            ROOT / "crates" / "tsz-emitter" / "src",
-            ROOT / "crates" / "tsz-lsp" / "src",
-            ROOT / "crates" / "tsz-cli" / "src",
-            ROOT / "crates" / "tsz-wasm" / "src",
-        ],
-        (),
-        16,
-    ),
-]
-
-# Pin the producer-side compatibility surface that still re-exports solver
-# computation/construction APIs from the crate root. The zero wildcard guard
-# below prevents broad `pub use module::*` growth; this count makes explicit
-# root re-export growth visible too.
-#
-# Each entry:
-#   (description, file_path, root_module_prefixes, max_reexports).
-ROOT_SOLVER_EXPLICIT_REEXPORT_COUNT_CHECKS = [
-    (
-        "Solver API boundary: flat root explicit computation re-exports (#8204)",
-        ROOT / "crates" / "tsz-solver" / "src" / "lib.rs",
-        (
-            "caches",
-            "canonicalize",
-            "classes",
-            "contextual",
-            "evaluation",
-            "instantiation",
-            "intern",
-            "narrowing",
-            "objects",
-            "operations",
-            "relations",
-            "widening",
-        ),
-        0,
-    ),
-]
-
-# Pin direct checker call sites into `query_boundaries::common`, the broad
-# compatibility/quarantine barrel tracked by #8225. Existing sites are
-# tolerated as migration debt; new checker code should prefer a narrower
-# request-shaped boundary module, or intentionally bump this cap.
-#
-# Each entry:
-#   (description, search_roots, exclude_path_prefixes, max_references).
-QUERY_BOUNDARY_COMMON_REFERENCE_COUNT_CHECKS = [
-    (
-        "Checker query boundary: direct common quarantine references outside query_boundaries (#8225)",
-        [ROOT / "crates" / "tsz-checker" / "src"],
-        ("crates/tsz-checker/src/query_boundaries/",),
-        # Bumped by 2 for the deferred-conditional diagnostic-display fix
-        # (`is_conditional_type` guards in the assignment-target display path,
-        # matching the existing direct-call pattern in type_display.rs).
-        #
-        # Ratcheted down by 5 after literal alias / literal widening
-        # diagnostic display probes moved through query_boundaries::diagnostics.
-        #
-        # Ratcheted down by 14 after branch refresh removed stale direct
-        # common references.
-        #
-        # Ratcheted down by 1 during the #9281 current-main refresh after
-        # the split guard tests caught slack in the live count.
-        #
-        # Ratcheted down by 1 after the interface heritage `this`-type helper
-        # moved to `query_boundaries::type_predicates`.
-        #
-        # Ratcheted down by 8 after rebasing on main removed additional direct
-        # common references.
-        #
-        # Ratcheted down to the live merged Application-source refresh count
-        # after current main reduced the quarantine surface.
-        #
-        # Ratcheted down to the live merged count after #10311 and #10359
-        # narrowed checker-side direct common references; removal condition
-        # remains #8225 narrowing this quarantine.
-        #
-        # Ratcheted down after arch-smoke caught current stacked-branch slack.
-        #
-        # Bumped by 3 for the intersection-target missing-property elaboration
-        # (#11480): `render_missing_property` adds two `is_intersection_type`
-        # guards picking the matched intersection plus an
-        # `intersection_members` + `find_property_by_str` member lookup, all
-        # matching the existing direct-call pattern already used throughout this
-        # file. The helpers are only exposed via `query_boundaries::common`.
-        #
-        # Bumped by 8 for the TS2352 constrained-type-parameter assertion
-        # rule (#10676) — `assertion_source_fits_constrained_type_param` and
-        # its recursive helper resolve / decompose the constraint via
-        # `query_boundaries::common` helpers (`type_param_info`,
-        # `union_members`, `intersection_members`, `find_property_in_object`,
-        # `object_shape_for_type`, `types_are_comparable_for_assertion`).
-        # All routes are existing request-shaped boundaries — no new
-        # quarantine entry — so the count rises by the expected delta.
-        #
-        # Bumped by 1 for the concrete indexed-access TS2536→TS2339 parity fix:
-        # the missing-literal-key guard in `check_indexed_access_type` adds one
-        # `contains_type_parameters` concreteness check (the literal key name is
-        # derived through `query_boundaries::type_computation::access`, not the
-        # `common` barrel). It is an existing request-shaped helper already used
-        # throughout this file — no new quarantine entry.
-        # Ratcheted 3278→3202 after current arch-smoke caught live-count slack.
-        #
-        # Ratcheted 3202→3197 after generic predicate-target instantiation
-        # moved five flow/narrowing substitution queries through
-        # query_boundaries::flow_analysis instead of the broad common barrel.
-        #
-        # Ratcheted 3197→3191 after assignment flow operator/property/literal
-        # queries moved through query_boundaries::flow_analysis.
-        #
-        # Ratcheted 3191→3174 after assignment numeric display shape queries
-        # moved through query_boundaries::diagnostics.
-        #
-        # Ratcheted 3174→3163 after missing-property display shape queries
-        # moved through query_boundaries::diagnostics.
-        #
-        # Ratcheted 3163→3155 after arch-smoke caught current-main slack in the
-        # live direct-reference count.
-        #
-        # Ratcheted 3155→3040 after the #8204 module-path ratchet PR's
-        # arch-smoke run caught current-main slack in the live count.
-        3040,
-    ),
-]
-
-# Pin root-level lint allowance entries in the query-boundary module map. #8225
-# tracks turning this layer from migration quarantine into narrower APIs, and
-# broad module-level allowances are part of that quarantine debt. The cap should
-# ratchet down as modules no longer need blanket suppressions.
-QUERY_BOUNDARY_MODULE_ALLOWANCE_COUNT_CHECKS = [
-    (
-        "Checker query boundary: module-level lint allowances must not grow (#8225)",
-        ROOT / "crates" / "tsz-checker" / "src" / "query_boundaries" / "mod.rs",
-        0,
-    ),
-]
-
-WORKSPACE_CLIPPY_ALLOW_COUNT_CHECKS = [
-    (
-        "Workspace Clippy suppressions must not grow (#9446)",
-        [ROOT / "crates"],
-        # Bumped 10 -> 11 for the JSX special-attribute display split: the new
-        # `check_assignable_or_report_at_with_display_types_and_options` helper
-        # legitimately takes eight explicit parameters (real + display source /
-        # target, two node indices, two display-policy flags), which with the
-        # receiver exceeds the workspace `too-many-arguments-threshold = 8`, so
-        # a single `#[allow(clippy::too_many_arguments)]` is required.
-        11,
-    ),
-]
-
-SNAPSHOT_ROLLBACK_FILE_COUNT_CHECKS = [
-    (
-        "Checker speculation boundary: snapshot-rollback call sites outside speculation.rs (architecture health metric 5)",
-        [ROOT / "crates" / "tsz-checker" / "src"],
-        ("crates/tsz-checker/src/context/speculation.rs",),
-        6,
-    ),
-]
-
-# Pin architecture health metric 6 ("Speculation APIs with surprising
-# non-RAII behavior") in `docs/plan/ROADMAP.md`.
-#
-# After PR #1213 renamed `DiagnosticSpeculationGuard → DiagnosticSpeculationSnapshot`
-# the speculation surface no longer carries `…Guard` types whose name
-# implies RAII rollback-on-drop while the implementation is implicit-commit.
-# This guard pins the rename: any new `pub(crate) struct …Guard` on the
-# speculation surface re-introduces the same ambiguity and must update
-# the cap (deliberately) or use a `…Snapshot` name (preferred). The
-# scan looks at the speculation file directly so the check is local
-# and cheap.
-#
-# Each entry: (description, file_path, max_guard_struct_count).
-SPECULATION_GUARD_NAME_CHECKS = [
-    (
-        "Checker speculation boundary: number of `…Guard` structs in speculation.rs (architecture health metric 6)",
-        ROOT / "crates" / "tsz-checker" / "src" / "context" / "speculation.rs",
-        0,
-    ),
-]
-
-DEBUG_PRINT_REPORT_PATH = ROOT / "scripts" / "perf" / "debug-print-report.py"
-DEBUG_PRINT_MACRO_CHECKS = [
-    (
-        "Performance boundary: compiler-internal debug print macros (Track 10)",
-        ROOT,
-        (
-            "crates/tsz-binder/src",
-            "crates/tsz-checker/src",
-            "crates/tsz-common/src",
-            "crates/tsz-core/src",
-            "crates/tsz-emitter/src",
-            "crates/tsz-lowering/src",
-            "crates/tsz-parser/src",
-            "crates/tsz-scanner/src",
-            "crates/tsz-solver/src",
-        ),
-    ),
-]
-
-# Pin Track 10's diagnostic-debt ratchets in the shared architecture guard.
-# These are count metrics, not new semantic bans: the current baselines still
-# contain legacy fingerprint rewrites, source-text snippets, and rendered-type
-# decisions. Any new line must bump the cap intentionally; cleanup PRs should
-# lower the cap in the same diff.
-#
-# Each entry: (description, search_roots, pattern, max_lines).
-REGEX_LINE_COUNT_CHECKS = [
-    (
-        "Checker diagnostic boundary: post-check rewrite_*_fingerprints functions (Track 10)",
-        [ROOT / "crates" / "tsz-checker" / "src"],
-        re.compile(r"^\s*fn\s+rewrite_\w+_fingerprints\s*\("),
-        6,
-    ),
-    (
-        "Checker diagnostic boundary: source_text.contains decisions (Track 10)",
-        [ROOT / "crates" / "tsz-checker" / "src"],
-        re.compile(r"\bsource_text\.contains\s*\("),
-        23,
-    ),
-    (
-        "Checker diagnostic boundary: file-name/path substring decisions (Track 10)",
-        [ROOT / "crates" / "tsz-checker" / "src"],
-        re.compile(r"\b(?:\w+\.)?file_name\.contains\s*\(|\bsource_path\.contains\s*\("),
-        1,
-    ),
-    (
-        "Checker diagnostic boundary: rendered type strings as semantic input (Track 10)",
-        [ROOT / "crates" / "tsz-checker" / "src"],
-        re.compile(
-            r"\bformat_type(?:_diagnostic)?\s*\([^\n]*"
-            r"(?:\.contains\s*\(|\.starts_with\s*\(|\.ends_with\s*\(|\.as_str\s*\(\))"
-        ),
-        0,
-    ),
-    (
-        "Checker diagnostic boundary: rendered message predicates (Track 10)",
-        [
-            ROOT / "crates" / "tsz-checker" / "src" / "checkers" / "jsx",
-            ROOT / "crates" / "tsz-checker" / "src" / "checkers" / "call_checker",
-            ROOT / "crates" / "tsz-checker" / "src" / "types" / "type_checking",
-        ],
-        re.compile(
-            r"\b(?:display|source_display|target_display|stripped_display|"
-            r"diagnostic\.message_text|raw|evaluated)"
-            r"\.(?:contains|starts_with|ends_with|as_str)\s*\("
-        ),
-        14,
-    ),
-    (
-        "Emitter boundary: source_text contains recovery decisions (Track 9/10)",
-        [ROOT / "crates" / "tsz-emitter" / "src"],
-        re.compile(r"\bsource_text(?:\[[^\n\]]+\])?\.contains\s*\("),
-        0,
-    ),
-    (
-        "Emitter boundary: recovered variable typeof tails use parser facts (#8276)",
-        [
-            ROOT
-            / "crates"
-            / "tsz-emitter"
-            / "src"
-            / "emitter"
-            / "statements"
-            / "recovered_variable_statement.rs"
-        ],
-        re.compile(
-            r"\b(?:find_source_pattern_outside_quoted_text|find_matching_source_paren|skip_quoted_source_text)\b"
-        ),
-        0,
-    ),
-    (
-        "Solver API boundary: flat root wildcard compatibility re-exports (#8204)",
-        [ROOT / "crates" / "tsz-solver" / "src" / "lib.rs"],
-        re.compile(r"^pub use (?:[A-Za-z_][A-Za-z0-9_]*::)+\*;"),
-        0,
-    ),
-    (
-        "Solver API boundary: root judge convenience re-export (#8204)",
-        [ROOT / "crates" / "tsz-solver" / "src" / "lib.rs"],
-        re.compile(r"^\s*pub\s+mod\s+judge\s*\{"),
-        0,
-    ),
-    (
-        "Solver relation boundary: legacy relation flag bridge surface (#8207)",
-        [ROOT / "crates" / "tsz-solver" / "src"],
-        re.compile(
-            r"\b(?:from_checker_flags_u16|from_legacy_u8|to_legacy_u8|"
-            r"subtype_cache_config_from_legacy_flags|"
-            r"assignability_cache_config_from_legacy_flags)\b"
-        ),
-        0,
-    ),
-    (
-        "Solver relation boundary: RelationPolicy must not expose packed flags (#8207)",
-        [ROOT / "crates" / "tsz-solver" / "src" / "relations" / "relation_queries.rs"],
-        re.compile(r"\bfn\s+legacy_packed_flags\s*\([^)]*\)\s*->\s*u16\b"),
-        0,
-    ),
-    (
-        "Checker relation boundary: raw diagnostic assignability predicates (#8227)",
-        [
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "assignability"
-            / "assignability_diagnostics.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "assignability"
-            / "assignment_checker"
-            / "destructuring.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "assignability"
-            / "assignment_checker"
-            / "assignment_ops.rs",
-            ROOT / "crates" / "tsz-checker" / "src" / "assignability" / "nullish_error_targets.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "assignability"
-            / "polymorphic_this_diagnostics.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "state"
-            / "state_checking_members"
-            / "decorator_signature_checks.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "state"
-            / "state_checking_members"
-            / "index_signature_checks.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "state"
-            / "state_checking_members"
-            / "index_signature_key_helpers.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "state"
-            / "state_checking_members"
-            / "overload_compatibility.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "state"
-            / "state_checking_members"
-            / "statement_helpers.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "state"
-            / "state_checking"
-            / "mapped_object_literals.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "state"
-            / "state_checking"
-            / "class.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "classes"
-            / "class_checker_compat.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "classes"
-            / "class_checker_compat_overloads.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "classes"
-            / "class_implements_checker"
-            / "core.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "classes"
-            / "class_implements_checker"
-            / "jsdoc_heritage.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "classes"
-            / "interface_heritage_index_compat.rs",
-            ROOT / "crates" / "tsz-checker" / "src" / "classes" / "interface_heritage_index.rs",
-            ROOT / "crates" / "tsz-checker" / "src" / "types" / "computation" / "access.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "types"
-            / "computation"
-            / "access_helpers.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "types"
-            / "type_checking"
-            / "core_statement_checks.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "types"
-            / "type_checking"
-            / "indexed_access.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "types"
-            / "type_checking"
-            / "indexed_access"
-            / "indexed_access_helpers.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "types"
-            / "type_checking"
-            / "indexed_access"
-            / "mapped_key_check.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "types"
-            / "property_access_type"
-            / "helpers.rs",
-            ROOT / "crates" / "tsz-checker" / "src" / "jsdoc" / "lookup.rs",
-            ROOT / "crates" / "tsz-checker" / "src" / "declarations" / "dynamic_import_checker.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "declarations"
-            / "import"
-            / "declaration.rs",
-            ROOT / "crates" / "tsz-checker" / "src" / "error_reporter",
-            ROOT / "crates" / "tsz-checker" / "src" / "checkers" / "call_context.rs",
-            ROOT / "crates" / "tsz-checker" / "src" / "checkers" / "promise_checker.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "checkers"
-            / "generic_checker"
-            / "explicit_alias_constraint_helpers.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "checkers"
-            / "generic_checker"
-            / "array_like_constraint_helpers.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "checkers"
-            / "generic_checker"
-            / "merged_interface_constraints.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "checkers"
-            / "generic_checker"
-            / "constraint_syntax_instantiation.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "checkers"
-            / "generic_checker"
-            / "infer_conditional_constraints.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "checkers"
-            / "generic_checker"
-            / "conditional_constraint_helpers.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "checkers"
-            / "generic_checker"
-            / "mapped_constraint_helpers.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "checkers"
-            / "generic_checker"
-            / "mod.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "checkers"
-            / "generic_checker"
-            / "recursive_heritage_constraint.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "checkers"
-            / "generic_checker"
-            / "union_constraint_helpers.rs",
-            ROOT / "crates" / "tsz-checker" / "src" / "checkers" / "iterable_checker.rs",
-            ROOT / "crates" / "tsz-checker" / "src" / "checkers" / "parameter_checker.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "state"
-            / "variable_checking"
-            / "core.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "state"
-            / "variable_checking"
-            / "initializer_policy.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "state"
-            / "variable_checking"
-            / "for_loop.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "checkers"
-            / "call_checker"
-            / "diagnostics.rs",
-            ROOT / "crates" / "tsz-checker" / "src" / "checkers" / "jsx",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "types"
-            / "computation"
-            / "call"
-            / "nominal_lib_object_callbacks.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "types"
-            / "computation"
-            / "call_finalize.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "types"
-            / "computation"
-            / "call_result.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "types"
-            / "computation"
-            / "call"
-            / "tail_helpers.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "state"
-            / "type_analysis"
-            / "computed_helpers_private.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "state"
-            / "type_analysis"
-            / "core.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "state"
-            / "type_analysis"
-            / "type_param_defaults.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "state"
-            / "state_checking"
-            / "property.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "checkers"
-            / "generic_checker"
-            / "constraint_indexed_access_helpers.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "checkers"
-            / "generic_checker"
-            / "constraint_validation.rs",
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "types"
-            / "computation"
-            / "complex_contextual_new.rs",
-        ],
-        re.compile(
-            r"\b(?:self|self\.ctx\.types|self\.interner)"
-            r"\.is_assignable_to(?:_[A-Za-z0-9_]+)?\s*\("
-        ),
-        0,
-    ),
-    (
-        "Checker residency boundary: with_parent_cache_attributed migration callsites (Track 10)",
-        [ROOT / "crates" / "tsz-checker" / "src"],
-        re.compile(
-            r"^(?!\s*(?:pub(?:\([^)]*\))?\s+)?fn\b)"
-            r".*\bwith_parent_cache_attributed\s*\("
-        ),
-        28,
-    ),
-    (
-        "Checker residency boundary: copy_symbol_file_targets_to_attributed migration callsites (Track 10)",
-        [ROOT / "crates" / "tsz-checker" / "src"],
-        re.compile(
-            r"^(?!\s*(?:pub(?:\([^)]*\))?\s+)?fn\b)"
-            r".*\bcopy_symbol_file_targets_to_attributed\s*\("
-        ),
-        18,
-    ),
-    (
-        "Checker relation boundary: diagnostic-local RelationRequest constructors (#8227)",
-        [
-            ROOT
-            / "crates"
-            / "tsz-checker"
-            / "src"
-            / "assignability"
-            / "assignability_diagnostics.rs",
-            ROOT / "crates" / "tsz-checker" / "src" / "error_reporter",
-            ROOT / "crates" / "tsz-checker" / "src" / "checkers" / "jsx",
-        ],
-        re.compile(r"\bRelationRequest::[A-Za-z_][A-Za-z0-9_]*\s*\("),
-        0,
-    ),
-    (
-        "Solver relation boundary: legacy packed relation flag bridges (#8207)",
-        [ROOT / "crates" / "tsz-solver" / "src"],
-        re.compile(
-            r'^(?:[^"\n]|"[^"\n]*")*?'
-            r"\b(?:subtype_cache_config_from_legacy_flags\s*\(|"
-            r"assignability_cache_config_from_legacy_flags\s*\(|"
-            r"from_checker_flags_u16\s*\(|from_legacy_u8\s*\(|to_legacy_u8\s*\(|"
-            r"RelationCacheKey::(?:subtype|assignability)\s*\(|"
-            r"RelationFlags::from_bits_truncate\s*\(|"
-            r"CachedAnyMode::from_legacy_u8\s*\()"
-        ),
-        0,
-    ),
-    (
-        "Solver relation boundary: relation engines avoid packed apply_flags (#8207)",
-        [ROOT / "crates" / "tsz-solver" / "src" / "relations"],
-        re.compile(
-            r"\bfn\s+apply_flags\s*\([^)]*\bflags\s*:\s*u16"
-            r"|\.\s*apply_flags\s*\(\s*policy\.flags\s*\)"
-        ),
-        0,
-    ),
-    (
-        "Solver relation boundary: legacy flag decoder avoids cache-key constants (#8207)",
-        [ROOT / "crates" / "tsz-solver" / "src" / "relations" / "relation_queries.rs"],
-        re.compile(r"\bRelationCacheKey::FLAG_[A-Z0-9_]+\b"),
-        0,
-    ),
-    (
-        "Solver relation boundary: legacy RelationPolicy::from_flags calls stay at boundary (#8207)",
-        [ROOT / "crates" / "tsz-solver" / "src"],
-        re.compile(
-            r'^\s*(?!//)(?:[^"\n]|"[^"\n]*")*?'
-            r"\bRelationPolicy::from_flags\s*\("
-        ),
-        0,
-    ),
-    (
-        "Solver relation boundary: query cache uses relation facade (#8207)",
-        [ROOT / "crates" / "tsz-solver" / "src" / "caches" / "query_cache.rs"],
-        re.compile(r"\b(?:configured_compat_checker|configured_subtype_checker)\s*\("),
-        0,
-    ),
-    (
-        "Solver relation boundary: query cache trace labels use typed policy names (#8207)",
-        [ROOT / "crates" / "tsz-solver" / "src" / "caches" / "query_cache.rs"],
-        re.compile(r'"is_(?:subtype_of|assignable_to)_with_flags"'),
-        0,
-    ),
-    (
-        "Solver relation boundary: query cache legacy flag overrides (#8207)",
-        [ROOT / "crates" / "tsz-solver" / "src" / "caches" / "query_cache.rs"],
-        re.compile(r"\bfn\s+is_(?:subtype_of|assignable_to)_with_flags\s*\("),
-        0,
-    ),
-    (
-        "Solver relation boundary: query database legacy flag methods (#8207)",
-        [
-            ROOT / "crates" / "tsz-solver" / "src" / "caches" / "db.rs",
-            ROOT / "crates" / "tsz-solver" / "src" / "caches" / "query_cache.rs",
-        ],
-        re.compile(r"\bfn\s+is_(?:subtype_of|assignable_to)_with_flags\s*\("),
-        0,
-    ),
-]
-
-# Track 10 performance guardrail: branch-local `visited.clone()` traversal
-# clones are a known scale-cliff risk for graph predicates.  Existing sites are
-# pinned by file plus statement text so normal line movement does not churn the
-# guard, while new clone sites must either replace an existing one with a
-# memoized/worklist traversal or extend this allowlist intentionally.
-BRANCH_LOCAL_VISITED_CLONE_CHECKS = [
-    (
-        "Performance boundary: branch-local visited.clone() graph traversal sites (Track 10)",
-        [
-            ROOT / "crates" / "tsz-checker" / "src",
-            ROOT / "crates" / "tsz-lsp" / "src",
-            ROOT / "crates" / "tsz-solver" / "src",
-        ],
-        (
-            (
-                "crates/tsz-checker/src/state/type_environment/lazy.rs",
-                "let mut branch_visited = visited.clone();",
-            ),
-            (
-                "crates/tsz-checker/src/state/type_resolution/module.rs",
-                "let mut inner_visited = visited.clone();",
-            ),
-            (
-                "crates/tsz-checker/src/types/queries/type_only.rs",
-                "let mut exists_visited = visited.clone();",
-            ),
-            (
-                "crates/tsz-checker/src/types/queries/type_only.rs",
-                "let mut type_only_visited = visited.clone();",
-            ),
-            (
-                "crates/tsz-solver/src/evaluation/evaluate_rules/infer_pattern.rs",
-                "let mut alias_visited = visited.clone();",
-            ),
-            (
-                "crates/tsz-solver/src/evaluation/evaluate_rules/infer_pattern_object_helpers.rs",
-                "let mut alias_visited = visited.clone();",
-            ),
-        ),
-    ),
-]
-
-# Pin the count of LSP feature-dispatch methods in
-# `crates/tsz-lsp/src/project/features.rs` (architecture health metric 7
-# in `docs/plan/ROADMAP.md` — "LSP/WASM semantic features implemented
-# outside the compiler service layer").
-#
-# Every `pub fn` on `Project` whose name starts with one of `get_`,
-# `provide_`, `prepare_`, `handle_`, `on_`, `find_`, or `resolve_` is an
-# LSP feature dispatched directly from `Project` rather than through a
-# service-trait abstraction. Workstream 6 ("LSP And WASM As Service
-# Clients") exit criterion 3 is that "LSP request handling mostly maps
-# protocol inputs to service queries and service outputs to protocol
-# DTOs"; the raw count tracks how far the live code is from that
-# state. Each new feature dispatch must bump the cap with a roadmap
-# entry; consolidation onto a service trait shows up as a cap reduction
-# in the same diff.
-#
-# Each entry: (description, file_path, max_methods).
-LSP_FEATURE_METHOD_COUNT_CHECKS = [
-    (
-        "LSP boundary: feature-dispatch method count in project/features.rs (architecture health metric 7)",
-        ROOT / "crates" / "tsz-lsp" / "src" / "project" / "features.rs",
-        32,
-    ),
-]
-
-PROJECT_DASHBOARD_ROW_CHECKS = [
-    (
-        "Project corpus dashboard: shared project row manifest must cover dashboard rows (Track 1)",
-        ROOT / "scripts" / "bench" / "project-rows.mjs",
-    ),
-]
-
-PROJECT_FIXTURE_SOURCE_CHECKS = [
-    (
-        "Project corpus fixtures: pinned rows must record fixture source refs (Track 1)",
-        ROOT / "scripts" / "bench" / "project-rows.mjs",
-        ROOT / "scripts" / "bench" / "project-fixtures.sh",
-    ),
-]
-
-PROJECT_INCLUSION_POLICY_CHECKS = [
-    (
-        "Project corpus inclusion: row manifest must match compile guard and benchmark rows (Track 1)",
-        ROOT / "scripts" / "bench" / "project-rows.mjs",
-        ROOT / "scripts" / "ci" / "project-compile-guard.sh",
-        ROOT / "scripts" / "bench" / "bench-vs-tsgo.sh",
-    ),
-]
-
-PROJECT_CONFIG_WRITER_CHECKS = [
-    (
-        "Project corpus config shape: shared rows must use shared config writers (Track 1)",
-        ROOT / "scripts" / "bench" / "project-fixtures.sh",
-        ROOT / "scripts" / "ci" / "project-compile-guard.sh",
-        ROOT / "scripts" / "bench" / "bench-vs-tsgo.sh",
-    ),
-]
-
-PROJECT_CONFIG_WRITERS = {
-    "utility-types-project": "tsz_write_utility_types_config",
-    "ts-toolbelt-project": "tsz_write_ts_toolbelt_config",
-    "ts-essentials-project": "tsz_write_ts_essentials_config",
-    "rxjs-project": "tsz_write_rxjs_config",
-    "type-fest-project": "tsz_write_type_fest_config",
-    "zod-project": "tsz_write_zod_config",
-    "kysely-project": "tsz_write_kysely_config",
-    "nextjs": "tsz_write_nextjs_config",
-}
-
-GENERATED_PROJECT_ROWS_WITHOUT_PINNED_SOURCE = {
-    "vite-vanilla-ts-app",
-    "nextjs-fresh-app",
-}
-
-COMPILE_GUARD_ONLY_PROJECT_ROWS = {
-    "type-challenges-solutions-project",
-}
-
-BENCHMARK_ONLY_PROJECT_ROWS = {
-    "nextjs",
-    "large-ts-repo",
-}
-
-EXCLUDE_DIRS = {".git", "target", "node_modules"}
-SOLVER_TYPEDATA_QUARANTINE_ALLOWLIST = {
-    "crates/tsz-solver/src/intern/mod.rs",
-    "crates/tsz-solver/src/intern/core/constructors.rs",
-    "crates/tsz-solver/src/intern/intersection.rs",
-    "crates/tsz-solver/src/intern/normalize.rs",
-    "crates/tsz-solver/src/intern/template.rs",
-}
